Split package into core and helper. Adjusted import statements in tests.
