+++ lib/python-beta/src/bailo/__init__.py
@@ -1,6 +1,6 @@
-"""Bailo"""
-from __future__ import annotations
-from bailo.bailo import *
-from bailo.enums import *
+"""Client for Bailo."""
 
-__version__ = "0.0.1"
+from __future__ import annotations
+from bailo.core import *
+
+__version__ = "0.1.0"
