+++ lib/python-beta/src/bailo/core/enums.py
