+++ lib/python-beta/tests/test_client.py
@@ -1,13 +1,7 @@
-#   ---------------------------------------------------------------------------------
-#   Copyright (c) Microsoft Corporation. All rights reserved.
-#   Licensed under the MIT License. See LICENSE in project root for information.
-#   ---------------------------------------------------------------------------------
-"""This is a sample python file for testing functions from the source code."""
 from __future__ import annotations
 import json
 
-from bailo import BailoClient
-from bailo.enums import ModelVisibility, SchemaKind
+from bailo import Client, ModelVisibility, SchemaKind
 
 mock_result = {"success": True}
 
@@ -15,7 +9,7 @@ mock_result = {"success": True}
 def test_post_model(requests_mock):
     requests_mock.post("https://example.com/api/v2/models", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.post_model(
         name="test",
         description="test",
@@ -28,7 +22,7 @@ def test_post_model(requests_mock):
 def test_get_models(requests_mock):
     requests_mock.get("https://example.com/api/v2/models/search?task=image_classification", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_models(
         task="image_classification"
     )
@@ -39,7 +33,7 @@ def test_get_models(requests_mock):
 def test_get_model(requests_mock):
     requests_mock.get("https://example.com/api/v2/model/test_id", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_model(
         model_id="test_id"
     )
@@ -50,7 +44,7 @@ def test_get_model(requests_mock):
 def test_patch_model(requests_mock):
     requests_mock.patch("https://example.com/api/v2/model/test_id", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.patch_model(
         model_id="test_id",
         name="test",
@@ -62,7 +56,7 @@ def test_patch_model(requests_mock):
 def test_get_model_card(requests_mock):
     requests_mock.get("https://example.com/api/v2/model/test_id/model-card/v1", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_model_card(
         model_id="test_id", version="v1"
     )
@@ -76,7 +70,7 @@ def test_put_model_card(requests_mock):
     x = {"test" : "object"}
     y = json.dumps(x)
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.put_model_card(
         model_id="test_id", metadata=y
     )
@@ -87,7 +81,7 @@ def test_put_model_card(requests_mock):
 def test_model_card_from_schema(requests_mock):
     requests_mock.post("https://example.com/api/v2/model/test_id/setup/from-schema", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.model_card_from_schema(
         model_id="test_id", schema_id='test_id'
     )
@@ -98,7 +92,7 @@ def test_model_card_from_schema(requests_mock):
 def test_post_release(requests_mock):
     requests_mock.post("https://example.com/api/v2/model/test_id/releases", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.post_release(
         model_id="test_id", 
         model_card_version=1, 
@@ -114,7 +108,7 @@ def test_post_release(requests_mock):
 def test_get_all_releases(requests_mock):
     requests_mock.get("https://example.com/api/v2/model/test_id/releases", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_all_releases(
         model_id="test_id",
     )
@@ -125,7 +119,7 @@ def test_get_all_releases(requests_mock):
 def test_get_release(requests_mock):
     requests_mock.get("https://example.com/api/v2/model/test_id/release/v1", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_release(
         model_id="test_id",
         release_version="v1",
@@ -137,7 +131,7 @@ def test_get_release(requests_mock):
 def test_delete_release(requests_mock):
     requests_mock.delete("https://example.com/api/v2/model/test_id/release/v1", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.delete_release(
         model_id="test_id",
         release_version="v1",
@@ -149,7 +143,7 @@ def test_delete_release(requests_mock):
 def test_get_files(requests_mock):
     requests_mock.get("https://example.com/api/v2/model/test_id/files", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_files(
         model_id="test_id",
     )
@@ -162,7 +156,7 @@ def test_simple_upload(requests_mock):
     data = 'TEST'
     data = bytes(data, 'utf-8')
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.simple_upload(
         model_id="test_id",
         name="test.txt",
@@ -178,7 +172,7 @@ def test_simple_upload(requests_mock):
 def test_delete_file(requests_mock):
     requests_mock.delete("https://example.com/api/v2/model/test_id/files/test_id", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.delete_file(
         model_id="test_id",
         file_id="test_id",
@@ -189,7 +183,7 @@ def test_delete_file(requests_mock):
 def test_get_all_schemas(requests_mock):
     requests_mock.get("https://example.com/api/v2/schemas?kind=model", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_all_schemas(
         kind=SchemaKind.Model
     )
@@ -199,7 +193,7 @@ def test_get_all_schemas(requests_mock):
 def test_get_schema(requests_mock):
     requests_mock.get("https://example.com/api/v2/schema/test_id", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_schema(
         schema_id="test_id"
     )
@@ -209,7 +203,7 @@ def test_get_schema(requests_mock):
 def test_post_schema(requests_mock):
     requests_mock.post("https://example.com/api/v2/schemas", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.post_schema(
         schema_id="test_id",
         name="test",
@@ -222,7 +216,7 @@ def test_post_schema(requests_mock):
 def test_get_reviews(requests_mock):
     requests_mock.get("https://example.com/api/v2/reviews?active=true", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_reviews(
         active=True,
     )
@@ -232,7 +226,7 @@ def test_get_reviews(requests_mock):
 def test_get_model_roles(requests_mock):
     requests_mock.get("https://example.com/api/v2/model/test_id/roles", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_model_roles(
         model_id="test_id",
     )
@@ -242,7 +236,7 @@ def test_get_model_roles(requests_mock):
 def test_get_model_user_roles(requests_mock):
     requests_mock.get("https://example.com/api/v2/model/test_id/roles/mine", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_model_user_roles(
         model_id="test_id",
     )
@@ -252,7 +246,7 @@ def test_get_model_user_roles(requests_mock):
 def test_post_team(requests_mock):
     requests_mock.post("https://example.com/api/v2/teams", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.post_team(
         team_id="test_id",
         name="test",
@@ -264,7 +258,7 @@ def test_post_team(requests_mock):
 def test_get_all_teams(requests_mock):
     requests_mock.get("https://example.com/api/v2/teams", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_all_teams()
 
     assert result == {"success": True}
@@ -272,7 +266,7 @@ def test_get_all_teams(requests_mock):
 def test_get_user_teams(requests_mock):
     requests_mock.get("https://example.com/api/v2/teams/mine", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_user_teams()
 
     assert result == {"success": True}
@@ -280,7 +274,7 @@ def test_get_user_teams(requests_mock):
 def test_get_team(requests_mock):
     requests_mock.get("https://example.com/api/v2/team/test_id", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.get_team(
         team_id="test_id",
     )
@@ -290,7 +284,7 @@ def test_get_team(requests_mock):
 def test_patch_team(requests_mock):
     requests_mock.patch("https://example.com/api/v2/team/test_id", json={"success": True})
 
-    client = BailoClient("https://example.com")
+    client = Client("https://example.com")
     result = client.patch_team(
         team_id="test_id",
         name="name",
